Atualizações e novas implementações do CRUD

+++ src/DevLearning/DevLearning.Api/Services/StudentService.cs
@@ -1,5 +1,6 @@
 ﻿using DevLearning.Api.Models;
-using DevLearning.Api.Models.Dtos;
+using DevLearning.Api.Models.Dtos.Student;
+using DevLearning.Api.Models.Dtos.StudentCourse;
 using DevLearning.Api.Repositories;
 
 namespace DevLearning.Api.Services
@@ -47,9 +48,13 @@ namespace DevLearning.Api.Services
 
         public async Task<List<StudentResponseDto>> GetAllStudentsAsync()
         {
+            var students = await _repository.GetAllStudentsAsync();
+            if (students is null)
+                throw new ArgumentException("Lista Vazia!");
+
             try
             {
-                return await _repository.GetAllStudentsAsync();
+                return students;
             }
             catch (Exception ex)
             {
@@ -59,6 +64,10 @@ namespace DevLearning.Api.Services
 
         public async Task<StudentResponseDto?> GetStudentByIdAsync(Guid id)
         {
+            var studentExist = await _repository.GetStudentByIdAsync(id);
+            if (studentExist is null)
+                throw new KeyNotFoundException("Estudante não encontrado!");
+
             try
             {
                 return await _repository.GetStudentByIdAsync(id);
@@ -71,12 +80,14 @@ namespace DevLearning.Api.Services
 
         public async Task UpdateStudentAsync(Guid id, UpdateStudentDto student)
         {
+            var studentExist = await _repository.GetStudentByIdAsync(id);
+            if (studentExist is null)
+                throw new KeyNotFoundException("Estudante não encontrado!");
+
             int number = 0;
             bool canConvert = int.TryParse(student.Phone, out number);
             if (canConvert == false)
-            {
                 throw new ArgumentException("Número de telefone inválido!");
-            }
 
             try
             {
@@ -90,6 +101,10 @@ namespace DevLearning.Api.Services
 
         public async Task DeleteStudentAsync(Guid id)
         {
+            var studentExist = await _repository.GetStudentByIdAsync(id);
+            if (studentExist is null)
+                throw new KeyNotFoundException("Estudante não encontrado!");
+
             try
             {
                 await _repository.DeleteStudentAsync(id);
@@ -100,5 +115,56 @@ namespace DevLearning.Api.Services
                 throw new Exception(ex.StackTrace);
             }
         }
+
+        public async Task CreateStudentCourseAsync(Guid courseId, Guid studentId, CreateStudentCourseDto studentCourse)
+        {
+
+            var studentExist = await _repository.GetStudentByIdAsync(studentId);
+            if (studentExist is null)
+                throw new KeyNotFoundException("Estudante não encontrado!");
+
+            try
+            {
+                var newStudentCourse = new StudentCourse
+                (
+                    courseId,
+                    studentId,
+                    studentCourse.Progress,
+                    studentCourse.Favorite
+                );
+
+                await _repository.CreateStudentCourseAsync(newStudentCourse);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.StackTrace);
+            }
+        }
+
+
+        public async Task<StudentCourse?> GetStudentCourseAsync(Guid courseId, Guid studentId)
+        {
+            
+        }
+        //TRATAR - NÃO EXISTE O ESTUDANTE SOZINHO (CADASTRADO)
+        //NÃO EXISTIR O GUID DO CURSO
+        //NÃO EXISTIR RELAÇÃO ENTRE ESTUDANTE E CURSO
+
+        public async Task UpdateStudentCourseProgressAsync(Guid id, UpdateStudentCourseDto student)
+        {
+            var studentExist = await _repository.GetStudentByIdAsync(id);
+            if (studentExist is null)
+                throw new KeyNotFoundException("Estudante não encontrado!");
+
+            try
+            {
+                await _repository.UpdateStudentCourseProgressAsync(id, student);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.StackTrace);
+            }
+        }
+
     }
 }
